+++ packages/plugin-packages/mashroom-portal-app-gallery-app/src/messages/messages-de.ts
@@ -1,0 +1,15 @@
+
+export default {
+    title: 'Mashroom Portal App Gallerie',
+    category: 'Kategorie',
+    tags: 'Tags',
+    homepage: 'Homepage',
+    permissions: 'Berechtigungen',
+    availableToEveryone: 'Verfügbar für alle',
+    roleRequired: 'Benötigte Rolle',
+    noScreenshot: 'Kein Screenshot',
+    notPermitted: 'Keine Bereichtigung',
+    filter: 'Filter',
+    tryOut: 'Ausprobieren',
+    errorLoading: 'Fehler beim Laden der Portal Apps'
+};