+++ packages/plugin-packages/mashroom-portal-app-gallery-app/src/messages/messages-en.ts
@@ -1,0 +1,15 @@
+
+export default {
+    title: 'Mashroom Portal App Gallery',
+    category: 'Category',
+    tags: 'Tags',
+    homepage: 'Homepage',
+    permissions: 'Permissions',
+    availableToEveryone: 'Available to Everyone',
+    roleRequired: 'Required Role',
+    noScreenshot: 'No Screenshot',
+    notPermitted: 'Not Permitted',
+    filter: 'Filter',
+    tryOut: 'Try Out',
+    errorLoading: 'Error loading Portal Apps'
+};