+++ packages/plugin-packages/mashroom-portal-app-gallery-app/src/modules.d.ts
@@ -1,0 +1,4 @@
+// Declare modules without existing TypeScript definitions
+
+declare module '*.svg';
+declare module '*.scss';